third_party/blink/web_tests/http/tests/devtools/elements/styles-1/case-sensitive-suggestions-expected.txt

Tests that text prompt suggestions' casing follows that of the user input.


Running: testForUpperCase

Running: testForLowerCase

Running: testForMixedCase
